Separate out ROOT dependent input tools to avoid loading root library if not necessary
